In this code, we've added the session_cookie method to the Auth class. This method takes a request object as an argument.
It first checks if request is None and returns None if so.
Then, it retrieves the name of the session cookie from the environment variable SESSION_NAME (defaulting to _my_session_id if not set).
Finally, it uses request.cookies.get(session_name) to retrieve the value of the session cookie from the request's cookies dictionary.
Now you can use this session_cookie method to access the session cookie value from a request object.